+++ lib/depject/index.js
@@ -111,8 +111,7 @@ module.exports = {
           post: require('./message/html/render/post.js'),
           vote: require('./message/html/render/vote.js'),
           'zzz-fallback': require('./message/html/render/zzz-fallback.js')
-        },
-        timestamp: require('./message/html/timestamp.js')
+        }
       },
       obs: {
         author: require('./message/obs/author.js'),
