+++ lib/depject/message/html/layout/default.js
@@ -2,6 +2,7 @@ const { h, computed, Value, when } = require('mutant')
 const nest = require('depnest')
 const ref = require('ssb-ref')
 const ExpanderHook = require('../../../../expander-hook')
+const timestamp = require('../../../../message/html/timestamp')
 
 exports.needs = nest({
   'profile.html.person': 'first',
@@ -13,7 +14,6 @@ exports.needs = nest({
     link: 'first',
     metas: 'first',
     actions: 'first',
-    timestamp: 'first',
     references: 'first',
     forks: 'first'
   },
@@ -144,7 +144,7 @@ exports.create = function (api) {
               msg.value.author === yourId ? [' ', h('span.you', {}, i18n('(you)'))] : null
             ]),
             h('div.meta', [
-              api.message.html.timestamp(msg), ' ',
+              timestamp(msg), ' ',
               replyInfo
             ])
           ])
