+++ lib/depject/message/html/layout/mini.js
@@ -2,6 +2,7 @@ const { h, computed, Value, when } = require('mutant')
 const nest = require('depnest')
 const ref = require('ssb-ref')
 const ExpanderHook = require('../../../../expander-hook')
+const timestamp = require('../../../../message/html/timestamp')
 
 exports.needs = nest({
   'profile.html.person': 'first',
@@ -12,7 +13,6 @@ exports.needs = nest({
     link: 'first',
     metas: 'first',
     actions: 'first',
-    timestamp: 'first',
     references: 'first',
     forks: 'first'
   },
@@ -123,7 +123,7 @@ exports.create = function (api) {
           ])
         ]),
         h('div.meta', [
-          h('strong', api.message.html.timestamp(msg)),
+          h('strong', timestamp(msg)),
           additionalMeta,
           api.message.html.metas(msg)
         ])
