+++ lib/depject/page/html/render/profile.js
@@ -2,6 +2,7 @@ const nest = require('depnest')
 const ref = require('ssb-ref')
 const { h, when, computed, map, send, dictToCollection, resolve, onceTrue, Value } = require('mutant')
 const displaySheet = require('../../../../sheet/display')
+const timestamp = require('../../../../message/html/timestamp')
 
 exports.needs = nest({
   'about.obs': {
@@ -15,7 +16,6 @@ exports.needs = nest({
   'blob.html.input': 'first',
   'message.async.publish': 'first',
   'message.html.markdown': 'first',
-  'message.html.timestamp': 'first',
   'about.html.image': 'first',
   'feed.html.rollup': 'first',
   'sbot.pull.resumeStream': 'first',
@@ -161,7 +161,7 @@ exports.create = function (api) {
         api.about.html.image(id),
         h('div.lastActivity',
           computed(lastActivity,
-            msg => `${i18n('Last activity')}: ${api.message.html.timestamp(msg, false)}`))]),
+            msg => `${i18n('Last activity')}: ${timestamp(msg, false)}`))]),
       h('div.main', [
         h('div.title', [
           h('h1', [name]),
